CV_DL/HW3-Instance-Segmentation/requirements.txt
numpy>=1.18.5
pandas==1.0.1
torch>=1.6.0
torchvision>=0.7.0
opencv-python==4.2.0
more-itertools==8.4.0
Pillow
matplotlib>=3.2.2
glob3
tqdm>=4.41.0
pycocotools==2.0.2